rewrote most of the audio management system, removed some unnecessary utils, amongst numerous other changes, see changelog

+++ src/main/java/com/yuuko/core/metrics/pathway/DiscordMetrics.java
@@ -4,13 +4,13 @@ import com.yuuko.core.Configuration;
 import net.dv8tion.jda.api.entities.Guild;
 
 import java.util.List;
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicInteger;
 
 public class DiscordMetrics {
     public int GUILD_COUNT = 0;
     public int USER_COUNT = 0;
-    public final AtomicLong GATEWAY_PING = new AtomicLong();
-    public final AtomicLong REST_PING = new AtomicLong();
+    public final AtomicInteger GATEWAY_PING = new AtomicInteger();
+    public final AtomicInteger REST_PING = new AtomicInteger();
 
     /**
      * Updates the Discord Metrics to current values.
@@ -31,7 +31,7 @@ public class DiscordMetrics {
      * Updates the bot's currently returned ping.
      */
     public void updatePing() {
-        GATEWAY_PING.set(Configuration.BOT.getJDA().getGatewayPing());
-        REST_PING.set(Configuration.BOT.getJDA().getRestPing().complete());
+        GATEWAY_PING.set((int)Configuration.BOT.getJDA().getGatewayPing());
+        REST_PING.set(Configuration.BOT.getJDA().getRestPing().complete().intValue());
     }
 }
